thermal: fix using custom or automatic warning levels, and set the get_functions
The constructor and applyConfig functions were using different semantics
for th->custom_levels. The dialog says "Automatic Warnings Levels",
which has the same meaning as "Not custom", hence rename the variable to
avoid future confusion, and use the same semantics in the two functions.
One could have used "auto_levels" instead, but since the configuration
uses the variable "CustomLevels", don't change that for a nice update
path.

Also, if automatic detection is used, the critical temperature as read
from the sensor needs to be updated. At the same time, the get_functions
may need updating, too, so do that as well.