taskbar: Fix infinite loop
Unlink a task from its class list properly, in case it is re-added. This
happens when you close windows of the same type in the order they were
created, and they for some reason add themselves again for a short time.
For example, LibreOffice writer 3.4.3 does this.

Reported in several places, e.g.
    - https://sourceforge.net/tracker/?func=detail&aid=3468295&group_id=180858&atid=894869
    - https://bugs.launchpad.net/ubuntu/+source/lxpanel/+bug/906825
    - https://sourceforge.net/mailarchive/forum.php?thread_name=13525c43570.-2435697889463439497.-8626413068332756300%40zoho.com&forum_name=lxde-list