taskbar: segfault fix - ID: 3573069
http://sourceforge.net/tracker/?func=detail&atid=894871&aid=3573069&group_id=180858

prevent segfault, zero out callback

This is my attempt to fix the segmentation fault with crashing function
flash_window_timeout reported here:
https://bugzilla.redhat.com/show_bug.cgi?id=587430

In various places in taskbar.c there is logic depending the timeout
callback pointer being 'active' when it is not zero. In one instance
(the one being address with the attached patch) the callback is
unregistered but not zeroed out.

This patch appears to prevent the segfault and also prevent excessive
flashing that Firefox exhibits in lxpanel.