change volumealsa plugin popup windwow from toplevel to popup
In volumealsa plugin, when creating a popup window, change window type from
toplevel to popup so window managers would ignore it. This way tiling window
managers do not tile it among other windows and do not make it a full screen
volume bar if that is the only open window.