Ship man pages in the GIT and tarball, not just xml files.
The XML files are not changed for years and doubtfully ever will.
There is no big reason to generate pages on each make or even ask
everyone to provide all the XML/XSL stuff for each GIT build.
Also it is possible that stylesheet may change over time and also
ROFF page is smaller than XML so it may be better to edit it instead.