Support HiDPI on GTK+ 3
For the most part, GDK 3 magically takes care of HiDPI support for us,
scaling between application-pixels and device-pixels.  But when we speak
directly to the X server (as when we set `_NET_WM_STRUT_PARTIAL`), this
translation doesn't happen for us.  So scale the numbers we send to the X
server by the GDK scale_factor.

Without this, lxpanel (with `--enable-gtk3`) properly renders at HiDPI
(assuming GDK_SCALE is configured correctly), but doesn't reserve
enough space at the edge of the screen.