Increase the size of the weather conditions icon in the panel
The icons provided by OpenWeatherMap occupy only a part (in the middle)
of the whole image area. When scaled down to panel size, they are barely
visible.

Cut off lines with no visible pixels at the top and bottom so that the
actual symbol remains larger after scaling.

Because cropping the image changes the aspect ratio, the new aspect
ratio must be saved and used when scaling the image.