Add option to thermal plugin to show first decimal place.
Some sensors can also supply the first decimal place instead of just
whole numbers. This commit adds support to use this additional precision
for display.
This necessitates changing the internal representation of temperatures from
degree celcius to a tenth of a degree.

This commit also adds logic to use an arbitrary file as sensor, which
works the same as reading a temperature from a file under /sys/class/thermal.
This was already possible before by exploiting the logic, as the checks would
treat a folder at an unexptected position as if it were under /proc/.
Now the support is explicit.