Correct icon-grid height in vertically oriented panels
When opening multiple windows, the Taskbar plugin does not resize
the taskbar buttons if they exceed the available space. With GTK+ 2,
they disappear behind the plugins below the taskbar, and with GTK +3,
everything below the taskbar slides off the screen instead.

Calculate child_width and child_height for GTK_ORIENTATION_VERTICAL
as well, which already fixes it for GTK +2.

For GTK+ 3 (additionally), do not set the minimum width to the same
value as the natural (i.e. current) width if the width is constrained
by the allocated space, but set it to the smallest possible value.