fix cpufreq applet
i submitted this fix on the sourceforge.net tracker several years ago,
and have used it ever since - it works splendidly.

in order to be able to use it, the user has to have the proper permissions,
which i achieve with a startup script doing this:

        for i in /sys/devices/system/cpu/cpu[0-9]*/cpufreq/scaling_governor /sys/devices/system/cpu/cpu[0-9]*/cpufreq/scaling_setspeed ; do
                chown root:cpufreq "$i"
                chmod 664 "$i"
        done

and having my user added to the cpufreq group.