netstat plugin: use copy of the linux kernel struct instead of headers
linux headers have the annoying property to override structs and types
defined in userland. in our case these caused build errors on musl libc.
we just copy the kernel struct definition that's needed here so we don't
have to include them. the kernel won't change the struct definition due
to its contract to not break userspace.