Fix the dynamic width of a panel with pager
Due to the need for improvement of the GTK 3 port and a change
in libwnck as of version 40, the reported size of the panel parent
class is incorrect and a horizontal panel ends up too short.

Use the size of the inner box of the panel as a workaround.

This closes github issue #86, reported by galtgendo.