Read desktop state from the window manager's _NET_SHOWING_DESKTOP hint
The desktop state (showing the desktop, i.e. all windows are hidden, or
windows are visible) is tracked with wc->toggle_state, a value that is
toggled each time the windows are minimized or restored.

However, if a window is clicked directly, there is no tracking with
wc->toggle_state, which means the variable will lose the information
about the current state of the window manager.

So, if available, get the current state by asking the window manager
for it.

Based on a patch by kurokawachan, github pull request #80, which has
been rewritten to be more straightforward.

This fixes github issue #79, reported by kurokawachan.