Restore disabled cpufreq plugin functionality
According to the comment used, this was apparently disabled because
someone thought that changing CPU governors requires root permissions,
but that can be easily changed using a rc script running on system
start.

Probably the most sensible setup to modify the governors from a
restricted user account is to add it to a dedicated group with
write access to the corresponding sysfs files.

For instance:

  for i in /sys/devices/system/cpu/cpu[0-9]*/cpufreq/scaling_governor \
           /sys/devices/system/cpu/cpu[0-9]*/cpufreq/scaling_setspeed ; do
          chown root:cpufreq "$i"
          chmod 664 "$i"
  done